My own implementation of the SQNL, but leaky :)
This one is of course not squashed to a value between -1 and 1, however, it is computationally efficient, and unlike SQNL, shouldn't have too many issues with the zero gradient at its' extrema. This function is designed to go in place of tanh, or something of the sort, and specifically be placed in hidden layers, not the output, since output typically will be between -1 to 1, rather than this. Hopefully this function of mine can help with performing faster training. The function of the derivative is a little "janky", it is not quite smooth, but this shouldn't cause too much of a problem :)